+++ src/Command/SchedulerTickCommand.php
@@ -11,7 +11,11 @@ use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
-#[AsCommand(name: 'core:scheduler:tick', description: 'Runs all scheduled jobs that are due at current minute.')]
+#[AsCommand(
+    name: 'core:scheduler:tick',
+    description: 'Runs all scheduled jobs that are due at current minute.',
+    aliases: ['app:scheduler:tick'],
+)]
 final class SchedulerTickCommand extends Command
 {
     public function __construct(private readonly SchedulerRunner $runner)