+++ src/Repository/Filters/DeletedFilter.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Core\Repository\Filters;
 
-use Doctrine\ORM\Mapping\ClassMetaData;
+use Doctrine\ORM\Mapping\ClassMetadata;
 use Doctrine\ORM\Query\Filter\SQLFilter;
 
 /**
@@ -15,10 +15,10 @@ class DeletedFilter extends SQLFilter
     /**
      * @param string $targetTableAlias
      */
-    public function addFilterConstraint(ClassMetaData $targetEntity, $targetTableAlias): string
+    public function addFilterConstraint(ClassMetadata $targetEntity, $targetTableAlias): string
     {
         if (in_array('Core\\Model\\Metadata\\DeleteInterface', class_implements($targetEntity->getName()), true)) {
-            return sprintf('%s.deleted = 0', $targetTableAlias);
+            return sprintf('%s.deleted = false', $targetTableAlias);
         }
 
         return '';